requirements.txt
oracledb==2.5.0
python-dotenv
sqlmodel
psycopg[binary]